+++ src/de/jost_net/JVerein/gui/control/AbstractAbrechnungControl.java
@@ -43,7 +43,6 @@ import de.jost_net.JVerein.gui.input.FormularInput;
 import de.jost_net.JVerein.gui.input.JVereinDateInput;
 import de.jost_net.JVerein.gui.menu.BugListMenu;
 import de.jost_net.JVerein.gui.parts.JVereinTablePart;
-import de.jost_net.JVerein.io.AbrechnungSEPA;
 import de.jost_net.JVerein.io.AbrechnungSEPAParam;
 import de.jost_net.JVerein.io.Bankarbeitstage;
 import de.jost_net.JVerein.keys.Abrechnungsausgabe;
@@ -590,7 +589,7 @@ public abstract class AbstractAbrechnungControl
             }
           }
           saveSettings();
-          new AbrechnungSEPA(getAbrechnungSEPAParam());
+          doStart();
           dialog.close();
         }
         catch (ApplicationException e)
@@ -1079,4 +1078,7 @@ public abstract class AbstractAbrechnungControl
       throws RemoteException, ApplicationException;
 
   protected abstract String checkInput();
+
+  protected abstract void doStart()
+      throws RemoteException, ApplicationException;
 }
