+++ src/de/jost_net/JVerein/gui/control/ArbeitseinsatzControl.java
@@ -16,71 +16,27 @@
  **********************************************************************/
 package de.jost_net.JVerein.gui.control;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.FileWriter;
 import java.rmi.RemoteException;
-import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Date;
-import java.util.HashMap;
-
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.widgets.Event;
-import org.eclipse.swt.widgets.FileDialog;
-import org.eclipse.swt.widgets.Listener;
-import org.supercsv.io.CsvMapWriter;
-import org.supercsv.io.ICsvMapWriter;
-import org.supercsv.prefs.CsvPreference;
-
-import com.itextpdf.text.BaseColor;
-import com.itextpdf.text.Element;
-
 import de.jost_net.JVerein.Einstellungen;
-import de.jost_net.JVerein.Einstellungen.Property;
 import de.jost_net.JVerein.gui.action.EditAction;
-import de.jost_net.JVerein.gui.input.ArbeitseinsatzUeberpruefungInput;
 import de.jost_net.JVerein.gui.menu.ArbeitseinsatzMenu;
 import de.jost_net.JVerein.gui.parts.ArbeitseinsatzPart;
-import de.jost_net.JVerein.gui.parts.ArbeitseinsatzUeberpruefungList;
 import de.jost_net.JVerein.gui.parts.JVereinTablePart;
 import de.jost_net.JVerein.gui.view.ArbeitseinsatzDetailView;
-import de.jost_net.JVerein.io.ArbeitseinsatzZeile;
-import de.jost_net.JVerein.io.Bankarbeitstage;
-import de.jost_net.JVerein.io.FileViewer;
-import de.jost_net.JVerein.io.Reporter;
-import de.jost_net.JVerein.keys.IntervallZusatzzahlung;
-import de.jost_net.JVerein.keys.VorlageTyp;
-import de.jost_net.JVerein.keys.Zahlungsweg;
 import de.jost_net.JVerein.rmi.Arbeitseinsatz;
-import de.jost_net.JVerein.rmi.Beitragsgruppe;
 import de.jost_net.JVerein.rmi.JVereinDBObject;
 import de.jost_net.JVerein.rmi.Mitglied;
-import de.jost_net.JVerein.rmi.Zusatzbetrag;
 import de.jost_net.JVerein.util.JVDateFormatTTMMJJJJ;
-import de.jost_net.JVerein.util.VorlageUtil;
-import de.willuhn.datasource.GenericIterator;
-import de.willuhn.datasource.GenericObject;
-import de.willuhn.datasource.pseudo.PseudoIterator;
 import de.willuhn.datasource.rmi.DBIterator;
 import de.willuhn.datasource.rmi.DBService;
 import de.willuhn.jameica.gui.AbstractView;
-import de.willuhn.jameica.gui.Action;
-import de.willuhn.jameica.gui.GUI;
 import de.willuhn.jameica.gui.Part;
 import de.willuhn.jameica.gui.formatter.CurrencyFormatter;
 import de.willuhn.jameica.gui.formatter.DateFormatter;
-import de.willuhn.jameica.gui.input.CheckboxInput;
-import de.willuhn.jameica.gui.input.DateInput;
-import de.willuhn.jameica.gui.input.SelectInput;
-import de.willuhn.jameica.gui.input.TextInput;
-import de.willuhn.jameica.gui.parts.Button;
-import de.willuhn.jameica.system.Application;
-import de.willuhn.jameica.system.BackgroundTask;
 import de.willuhn.jameica.system.Settings;
 import de.willuhn.logging.Logger;
 import de.willuhn.util.ApplicationException;
-import de.willuhn.util.ProgressMonitor;
 
 public class ArbeitseinsatzControl extends FilterControl implements Savable
 {
@@ -88,21 +44,7 @@ public class ArbeitseinsatzControl extends FilterControl implements Savable
 
   private Arbeitseinsatz aeins = null;
 
-  private ArbeitseinsatzUeberpruefungList arbeitseinsatzueberpruefungList;
-
-  private SelectInput suchjahr = null;
-
-  private ArbeitseinsatzUeberpruefungInput auswertungschluessel = null;
-
   private JVereinTablePart arbeitseinsatzList;
-
-  private TextInput buchungstext;
-
-  private SelectInput zahlungsweg;
-
-  private CheckboxInput mitgliedZahltSelbst;
-
-  private DateInput faelligkeit;
 
   public ArbeitseinsatzControl(AbstractView view)
   {
@@ -129,18 +71,6 @@ public class ArbeitseinsatzControl extends FilterControl implements Savable
     }
     part = new ArbeitseinsatzPart(getArbeitseinsatz(), true);
     return part;
-  }
-
-  public ArbeitseinsatzUeberpruefungInput getAuswertungSchluessel()
-      throws RemoteException
-  {
-    if (auswertungschluessel != null)
-    {
-      return auswertungschluessel;
-    }
-    auswertungschluessel = new ArbeitseinsatzUeberpruefungInput(1);
-    auswertungschluessel.addListener(new FilterListener());
-    return auswertungschluessel;
   }
 
   @Override
@@ -178,524 +108,6 @@ public class ArbeitseinsatzControl extends FilterControl implements Savable
       String fehler = "Fehler bei speichern des Arbeitseinsatzes";
       Logger.error(fehler, e);
       throw new ApplicationException(fehler, e);
-    }
-  }
-
-  public SelectInput getSuchJahr() throws RemoteException
-  {
-    if (suchjahr != null)
-    {
-      return suchjahr;
-    }
-    DBIterator<Arbeitseinsatz> list = Einstellungen.getDBService()
-        .createList(Arbeitseinsatz.class);
-    list.setOrder("ORDER BY datum");
-    Arbeitseinsatz ae = null;
-    Calendar von = Calendar.getInstance();
-    if (list.hasNext())
-    {
-      ae = list.next();
-      von.setTime(ae.getDatum());
-    }
-    Calendar bis = Calendar.getInstance();
-    ArrayList<Integer> jahre = new ArrayList<>();
-
-    for (int i = von.get(Calendar.YEAR); i <= bis.get(Calendar.YEAR); i++)
-    {
-      jahre.add(i);
-    }
-
-    suchjahr = new SelectInput(jahre, settings.getInt("jahr", jahre.get(0)));
-    // suchjahr.setPleaseChoose("Bitte auswählen");
-    suchjahr.setPreselected(settings.getInt("jahr", bis.get(Calendar.YEAR)));
-    suchjahr.addListener(new FilterListener());
-    return suchjahr;
-  }
-
-  public Button getPDFAusgabeButton()
-  {
-    Button b = new Button("PDF", new Action()
-    {
-
-      @Override
-      public void handleAction(Object context) throws ApplicationException
-      {
-        try
-        {
-          startePDFAuswertung();
-        }
-        catch (RemoteException e)
-        {
-          Logger.error(e.getMessage());
-          throw new ApplicationException(
-              "Fehler beim Start der PDF-Ausgabe der Arbeitseinsatzüberprüfung");
-        }
-      }
-    }, null, false, "file-pdf.png");
-    return b;
-  }
-
-  public Button getCSVAusgabeButton()
-  {
-    Button b = new Button("CSV", new Action()
-    {
-
-      @Override
-      public void handleAction(Object context) throws ApplicationException
-      {
-        try
-        {
-          starteCSVAuswertung();
-        }
-        catch (RemoteException e)
-        {
-          Logger.error(e.getMessage());
-          throw new ApplicationException(
-              "Fehler beim Start der CSV-Ausgabe der Arbeitseinsatzüberprüfung");
-        }
-      }
-    }, null, false, "xsd.png");
-    return b;
-  }
-
-  public Button getArbeitseinsatzAusgabeButton()
-  {
-    Button b = new Button("Zusatzbeträge generieren", new Action()
-    {
-
-      @Override
-      public void handleAction(Object context) throws ApplicationException
-      {
-        try
-        {
-          saveSettings();
-          starteArbeitseinsatzGenerierung();
-        }
-        catch (RemoteException e)
-        {
-          Logger.error(e.getMessage());
-          throw new ApplicationException(
-              "Fehler beim der Zusatzbetragsgenerierung");
-        }
-      }
-    }, null, false, "euro-sign.png");
-    return b;
-  }
-
-  private void startePDFAuswertung() throws RemoteException
-  {
-    FileDialog fd = new FileDialog(GUI.getShell(), SWT.SAVE);
-    fd.setText("Ausgabedatei wählen.");
-    String path = settings.getString("lastdir",
-        System.getProperty("user.home"));
-    if (path != null && path.length() > 0)
-    {
-      fd.setFilterPath(path);
-    }
-    fd.setFileName(VorlageUtil.getName(
-        VorlageTyp.AUSWERTUNG_ARBEITSEINSAETZE_DATEINAME, this) + ".pdf");
-    fd.setFilterExtensions(new String[] { "*.pdf" });
-
-    String s = fd.open();
-    if (s == null || s.length() == 0)
-    {
-      return;
-    }
-    if (!s.toLowerCase().endsWith(".pdf"))
-    {
-      s = s + ".pdf";
-    }
-    final File file = new File(s);
-    final GenericIterator<ArbeitseinsatzZeile> it = getIterator();
-    final String title = VorlageUtil
-        .getName(VorlageTyp.AUSWERTUNG_ARBEITSEINSAETZE_TITEL, this);
-    final String subtitle = VorlageUtil
-        .getName(VorlageTyp.AUSWERTUNG_ARBEITSEINSAETZE_SUBTITEL, this);
-    settings.setAttribute("lastdir", file.getParent());
-    BackgroundTask t = new BackgroundTask()
-    {
-
-      @Override
-      public void run(ProgressMonitor monitor) throws ApplicationException
-      {
-        try
-        {
-          FileOutputStream fos = new FileOutputStream(file);
-          Reporter reporter = new Reporter(fos, title, subtitle, it.size());
-          reporter.addHeaderColumn("Mitglied", Element.ALIGN_LEFT, 60,
-              BaseColor.LIGHT_GRAY);
-          reporter.addHeaderColumn("Sollstunden", Element.ALIGN_RIGHT, 30,
-              BaseColor.LIGHT_GRAY);
-          reporter.addHeaderColumn("Iststunden", Element.ALIGN_RIGHT, 30,
-              BaseColor.LIGHT_GRAY);
-          reporter.addHeaderColumn("Differenz", Element.ALIGN_RIGHT, 30,
-              BaseColor.LIGHT_GRAY);
-          reporter.addHeaderColumn("Stundensatz", Element.ALIGN_RIGHT, 30,
-              BaseColor.LIGHT_GRAY);
-          reporter.addHeaderColumn("Gesamtbetrag", Element.ALIGN_RIGHT, 30,
-              BaseColor.LIGHT_GRAY);
-          reporter.createHeader();
-          while (it.hasNext())
-          {
-            ArbeitseinsatzZeile z = (ArbeitseinsatzZeile) it.next();
-            if ((Boolean) Einstellungen
-                .getEinstellung(Property.MITGLIEDSNUMMERANZEIGEN))
-            {
-              reporter.addColumn((String) z.getAttribute("idnamevorname"),
-                  Element.ALIGN_LEFT);
-            }
-            else
-            {
-              reporter.addColumn((String) z.getAttribute("namevorname"),
-                  Element.ALIGN_LEFT);
-            }
-            reporter.addColumn((Double) z.getAttribute("soll"));
-            reporter.addColumn((Double) z.getAttribute("ist"));
-            reporter.addColumn((Double) z.getAttribute("differenz"));
-            reporter.addColumn((Double) z.getAttribute("stundensatz"));
-            reporter.addColumn((Double) z.getAttribute("gesamtbetrag"));
-          }
-          reporter.closeTable();
-          reporter.close();
-          fos.close();
-          GUI.getStatusBar().setSuccessText("Auswertung gestartet");
-          GUI.getCurrentView().reload();
-        }
-        catch (Exception e)
-        {
-          Logger.error("Fehler", e);
-          GUI.getStatusBar().setErrorText(e.getMessage());
-          throw new ApplicationException(e);
-        }
-        FileViewer.show(file);
-      }
-
-      @Override
-      public void interrupt()
-      {
-        //
-      }
-
-      @Override
-      public boolean isInterrupted()
-      {
-        return false;
-      }
-    };
-    Application.getController().start(t);
-  }
-
-  private void starteCSVAuswertung() throws RemoteException
-  {
-    FileDialog fd = new FileDialog(GUI.getShell(), SWT.SAVE);
-    fd.setText("Ausgabedatei wählen.");
-    String path = settings.getString("lastdir",
-        System.getProperty("user.home"));
-    if (path != null && path.length() > 0)
-    {
-      fd.setFilterPath(path);
-    }
-    fd.setFileName(VorlageUtil.getName(
-        VorlageTyp.AUSWERTUNG_ARBEITSEINSAETZE_DATEINAME, this) + ".csv");
-    fd.setFilterExtensions(new String[] { "*.csv" });
-
-    String s = fd.open();
-    if (s == null || s.length() == 0)
-    {
-      return;
-    }
-    if (!s.toLowerCase().endsWith(".csv"))
-    {
-      s = s + ".csv";
-    }
-    final File file = new File(s);
-    final GenericIterator<ArbeitseinsatzZeile> it = getIterator();
-    settings.setAttribute("lastdir", file.getParent());
-    BackgroundTask t = new BackgroundTask()
-    {
-
-      @Override
-      public void run(ProgressMonitor monitor) throws ApplicationException
-      {
-        try
-        {
-          ICsvMapWriter writer = new CsvMapWriter(new FileWriter(file),
-              CsvPreference.EXCEL_NORTH_EUROPE_PREFERENCE);
-
-          final String[] header = new String[] { "name", "vorname", "strasse",
-              "adressierungszusatz", "plz", "ort", "anrede", "telefonprivat",
-              "telefondienstlich", "handy", "email", "soll", "ist", "differenz",
-              "stundensatz", "gesamtbetrag" };
-          writer.writeHeader(header);
-          // set up some data to write
-          while (it.hasNext())
-          {
-            ArbeitseinsatzZeile z = (ArbeitseinsatzZeile) it.next();
-            final HashMap<String, ? super Object> data1 = new HashMap<>();
-            Mitglied m = (Mitglied) z.getAttribute("mitglied");
-            data1.put(header[0], m.getName());
-            data1.put(header[1], m.getVorname());
-            data1.put(header[2], m.getStrasse());
-            data1.put(header[3], m.getAdressierungszusatz());
-            data1.put(header[4], m.getPlz());
-            data1.put(header[5], m.getOrt());
-            data1.put(header[6], m.getAnrede());
-            data1.put(header[7], m.getTelefonprivat());
-            data1.put(header[8], m.getTelefondienstlich());
-            data1.put(header[9], m.getHandy());
-            data1.put(header[10], m.getEmail());
-            data1.put(header[11], z.getAttribute("soll"));
-            data1.put(header[12], z.getAttribute("ist"));
-            data1.put(header[13], z.getAttribute("differenz"));
-            data1.put(header[14], z.getAttribute("stundensatz"));
-            data1.put(header[15], z.getAttribute("gesamtbetrag"));
-            writer.write(data1, header);
-          }
-          writer.close();
-          GUI.getStatusBar().setSuccessText("Auswertung gestartet");
-          GUI.getCurrentView().reload();
-        }
-        catch (Exception e)
-        {
-          Logger.error("Fehler", e);
-          GUI.getStatusBar().setErrorText(e.getMessage());
-          throw new ApplicationException(e);
-        }
-        FileViewer.show(file);
-      }
-
-      @Override
-      public void interrupt()
-      {
-        //
-      }
-
-      @Override
-      public boolean isInterrupted()
-      {
-        return false;
-      }
-    };
-    Application.getController().start(t);
-
-  }
-
-  public TextInput getBuchungstext() throws RemoteException
-  {
-    if (buchungstext != null)
-    {
-      return buchungstext;
-    }
-    buchungstext = new TextInput(settings.getString("buchungstext", ""), 140);
-    buchungstext.setMandatory(true);
-    return buchungstext;
-  }
-
-  public SelectInput getZahlungsweg() throws RemoteException
-  {
-    if (zahlungsweg != null)
-    {
-      return zahlungsweg;
-    }
-    Zahlungsweg weg = new Zahlungsweg(Zahlungsweg.STANDARD);
-    String wegsetting = settings.getString("zahlungsweg", "");
-    if (wegsetting.length() > 0)
-    {
-      try
-      {
-        weg = new Zahlungsweg(Integer.valueOf(wegsetting));
-      }
-      catch (Exception e)
-      {
-        //
-      }
-    }
-    zahlungsweg = new SelectInput(Zahlungsweg.getArray(), weg);
-    zahlungsweg.setPleaseChoose("Standard");
-    return zahlungsweg;
-  }
-
-  public CheckboxInput getMitgliedzahltSelbst() throws RemoteException
-  {
-    if (mitgliedZahltSelbst != null)
-    {
-      return mitgliedZahltSelbst;
-    }
-    mitgliedZahltSelbst = new CheckboxInput(
-        settings.getBoolean("mitgliedzahltselbst", false));
-    mitgliedZahltSelbst
-        .setName(" *Falls ein abweichender Zahler konfiguriert ist.");
-    return mitgliedZahltSelbst;
-  }
-
-  public DateInput getFaelligkeit() throws RemoteException
-  {
-    if (faelligkeit != null)
-    {
-      return faelligkeit;
-    }
-    Calendar cal = Calendar.getInstance();
-    Bankarbeitstage bat = new Bankarbeitstage();
-    cal = bat.getCalendar(cal,
-        1 + (Integer) Einstellungen.getEinstellung(Property.SEPADATUMOFFSET));
-    this.faelligkeit = new DateInput(cal.getTime(), new JVDateFormatTTMMJJJJ());
-    faelligkeit.setMandatory(true);
-    return faelligkeit;
-  }
-
-  private void starteArbeitseinsatzGenerierung()
-      throws RemoteException, ApplicationException
-  {
-    final GenericIterator<ArbeitseinsatzZeile> it = getIterator();
-    String text = (String) getBuchungstext().getValue();
-    if (text == null || text.isEmpty())
-    {
-      throw new ApplicationException("Bitte Buchungstext eingeben.");
-    }
-    Date faelligkeit = (Date) getFaelligkeit().getValue();
-    if (faelligkeit == null)
-    {
-      throw new ApplicationException("Bitte Fälligkeit eingeben.");
-    }
-    Zahlungsweg zahlungsweg = (Zahlungsweg) getZahlungsweg().getValue();
-    Boolean mitgliedzahltselbst = (Boolean) getMitgliedzahltSelbst().getValue();
-
-    BackgroundTask t = new BackgroundTask()
-    {
-
-      @Override
-      public void run(ProgressMonitor monitor) throws ApplicationException
-      {
-        try
-        {
-          while (it.hasNext())
-          {
-            ArbeitseinsatzZeile z = (ArbeitseinsatzZeile) it.next();
-            Zusatzbetrag zb = (Zusatzbetrag) Einstellungen.getDBService()
-                .createObject(Zusatzbetrag.class, null);
-            Double betrag = (Double) z.getAttribute("gesamtbetrag");
-            betrag = betrag * -1;
-            zb.setBetrag(betrag);
-            zb.setBuchungstext(text);
-            zb.setFaelligkeit(faelligkeit);
-            zb.setStartdatum(faelligkeit);
-            zb.setIntervall(IntervallZusatzzahlung.KEIN);
-            zb.setMitglied(
-                Integer.valueOf((String) z.getAttribute("mitgliedid")));
-            zb.setZahlungsweg(zahlungsweg);
-            zb.setMitgliedzahltSelbst(mitgliedzahltselbst);
-            Mitglied m = (Mitglied) Einstellungen.getDBService().createObject(
-                Mitglied.class, (String) z.getAttribute("mitgliedid"));
-            Beitragsgruppe b = m.getBeitragsgruppe();
-            zb.setBuchungsart(b.getBuchungsart());
-            if ((Boolean) Einstellungen
-                .getEinstellung(Property.BUCHUNGSKLASSEINBUCHUNG))
-            {
-              zb.setBuchungsklasseId(b.getBuchungsklasseId());
-            }
-            zb.setSteuer(b.getSteuer());
-            zb.store();
-          }
-          GUI.getStatusBar().setSuccessText("Liste Arbeitseinsätze gestartet");
-          GUI.getCurrentView().reload();
-        }
-        catch (Exception e)
-        {
-          Logger.error("Fehler", e);
-          GUI.getStatusBar().setErrorText(e.getMessage());
-          throw new ApplicationException(e);
-        }
-
-      }
-
-      @Override
-      public void interrupt()
-      {
-        //
-      }
-
-      @Override
-      public boolean isInterrupted()
-      {
-        return false;
-      }
-    };
-    Application.getController().start(t);
-
-  }
-
-  private GenericIterator<ArbeitseinsatzZeile> getIterator()
-      throws RemoteException
-  {
-    ArrayList<ArbeitseinsatzZeile> zeile = arbeitseinsatzueberpruefungList
-        .getInfo();
-    @SuppressWarnings("unchecked")
-    GenericIterator<ArbeitseinsatzZeile> gi = PseudoIterator
-        .fromArray(zeile.toArray(new GenericObject[zeile.size()]));
-    return gi;
-  }
-
-  public Part getArbeitseinsatzUeberpruefungList() throws ApplicationException
-  {
-    try
-    {
-      settings.setAttribute("jahr", (Integer) getSuchJahr().getValue());
-      settings.setAttribute("schluessel",
-          (Integer) getAuswertungSchluessel().getValue());
-
-      if (arbeitseinsatzueberpruefungList == null)
-      {
-        arbeitseinsatzueberpruefungList = new ArbeitseinsatzUeberpruefungList(
-            null, (Integer) getSuchJahr().getValue(),
-            (Integer) getAuswertungSchluessel().getValue());
-      }
-      else
-      {
-        arbeitseinsatzueberpruefungList
-            .setJahr((Integer) getSuchJahr().getValue());
-        arbeitseinsatzueberpruefungList
-            .setSchluessel((Integer) getAuswertungSchluessel().getValue());
-        ArrayList<ArbeitseinsatzZeile> zeile = arbeitseinsatzueberpruefungList
-            .getInfo();
-        arbeitseinsatzueberpruefungList.removeAll();
-        for (ArbeitseinsatzZeile az : zeile)
-        {
-          arbeitseinsatzueberpruefungList.addItem(az);
-        }
-        arbeitseinsatzueberpruefungList.sort();
-      }
-    }
-    catch (RemoteException e)
-    {
-      Logger.error("Fehler", e);
-      throw new ApplicationException("Fehler aufgetreten", e);
-    }
-    return arbeitseinsatzueberpruefungList.getArbeitseinsatzUeberpruefungList();
-  }
-
-  private void refreshList()
-  {
-    try
-    {
-      getArbeitseinsatzUeberpruefungList();
-    }
-    catch (ApplicationException e1)
-    {
-      //
-    }
-  }
-
-  private class FilterListener implements Listener
-  {
-
-    @Override
-    public void handleEvent(Event event)
-    {
-      if (event.type != SWT.Selection && event.type != SWT.FocusOut)
-      {
-        return;
-      }
-      refreshList();
     }
   }
 
@@ -790,36 +202,5 @@ public class ArbeitseinsatzControl extends FilterControl implements Savable
     }
     arbeitseinsaetze.setOrder("ORDER by datum desc");
     return arbeitseinsaetze;
-  }
-
-  protected void saveSettings()
-  {
-    try
-    {
-      settings.setAttribute("buchungstext",
-          (String) getBuchungstext().getValue());
-      Zahlungsweg weg = (Zahlungsweg) getZahlungsweg().getValue();
-      if (weg != null)
-      {
-        settings.setAttribute("zahlungsweg", weg.getKey());
-      }
-      else
-      {
-        settings.setAttribute("zahlungsweg", "");
-      }
-      Boolean tmp = (Boolean) getMitgliedzahltSelbst().getValue();
-      if (tmp != null)
-      {
-        settings.setAttribute("mitgliedzahltselbst", tmp);
-      }
-      else
-      {
-        settings.setAttribute("mitgliedzahltselbst", "false");
-      }
-    }
-    catch (RemoteException re)
-    {
-      Logger.error("Fehler", re);
-    }
   }
 }
