+++ src/de/jost_net/JVerein/gui/control/GutschriftControl.java
@@ -955,4 +955,10 @@ public class GutschriftControl extends AbstractAbrechnungControl
   {
     return new AbrechnungSEPAParam(this, sepaVersion);
   }
+
+  @Override
+  public void doStart() throws ApplicationException
+  {
+    // wird hier nicht benutzt
+  }
 }
